Restored original expression transitions and added checkbox to disable them
Fix: Restored expression transitions from the original game.

They have been defined as dict transitions, to prevent the jarring flashing of the textbox in modern Ren'Py that was not present in the original game. This means the text isn't interrupted with each transition like it was in the original, but it's closer to the original than the alternative. It's also an improvement anyway, IMO.

Feat: Added a checkbox to disable the restored expression transitions.